learn3/pyproject.toml
[project]
name = "learn3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.7.5",
    "autogen-ext[openai]>=0.7.5",
    "python-dotenv>=1.2.2",
]
